so code clean up
