+++ src/styledGuide/DonationButton/index.js
@@ -1,23 +1,22 @@
 import React from "react";
-import { Grid } from "gymnast";
 
 /**TODO: replace with Diane's paypal donation button */
 
 export const DonationButton = () => {
   return (
-    <Grid size="fit" justify="center">
+    <div size="fit" justify="center">
       <form
         action="https://www.paypal.com/cgi-bin/webscr"
         method="post"
         target="_blank"
       >
-        <Grid>
+        <div>
           <input type="hidden" name="cmd" value="_s-xclick" />
-        </Grid>
-        <Grid>
+        </div>
+        <div>
           <input type="hidden" name="hosted_button_id" value="BD6MEHQFPGX2S" />
-        </Grid>
-        <Grid>
+        </div>
+        <div>
           <input
             type="image"
             src="https://www.paypalobjects.com/en_US/i/btn/btn_donate_SM.gif"
@@ -26,8 +25,8 @@ export const DonationButton = () => {
             title="PayPal - The safer, easier way to pay online!"
             alt="Donate with PayPal button"
           />
-        </Grid>
-        <Grid>
+        </div>
+        <div>
           <img
             alt="Donatation button"
             border="0"
@@ -35,8 +34,8 @@ export const DonationButton = () => {
             width="1"
             height="1"
           />
-        </Grid>
+        </div>
       </form>
-    </Grid>
+    </div>
   );
 };
