+++ src/styledGuide/Label/index.js
@@ -1,19 +1,8 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { Grid } from "gymnast";
 
-export const Label = ({
-  children,
-  size,
-  margin = "M M 0 M",
-  padding,
-  ...props
-}) => {
-  return (
-    <Grid size={size} margin={margin} padding={padding}>
-      <label {...props}>{children}</label>
-    </Grid>
-  );
+export const Label = ({ children, ...props }) => {
+  return <label {...props}>{children}</label>;
 };
 
 Label.propTypes = {
