+++ src/styledGuide/P/index.js
@@ -1,11 +1,8 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { useGrid } from "gymnast";
 
 export const P = ({ children, ...props }) => {
-  const [, allProps] = useGrid(props);
-
-  return <p {...allProps}>{children}</p>;
+  return <p {...props}>{children}</p>;
 };
 
 P.propTypes = {
